+++ lib/src/environment.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_environment.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 23c920bd76d38c4ccf2024a0740aeae9672143d0
+// Checksum: 70b94c80b7ecc1bab523f6bf69612b69269424a5
 //
 // ignore_for_file: unused_import
 
@@ -108,6 +108,9 @@ class Environment {
   /// this is not in a mixin, or if no content block was passed.
   UserDefinedCallable<Environment> get content => _content;
   UserDefinedCallable<Environment> _content;
+
+  /// Whether the environment is lexically at the root of the document.
+  bool get atRoot => _variables.length == 1;
 
   /// Whether the environment is lexically within a mixin.
   bool get inMixin => _inMixin;
@@ -361,7 +364,7 @@ class Environment {
       return;
     }
 
-    if (global || _variables.length == 1) {
+    if (global || atRoot) {
       // Don't set the index if there's already a variable with the given name,
       // since local accesses should still return the local variable.
       _variableIndices.putIfAbsent(name, () {